# docs/swagger.yaml
definitions:
  vo.AddSocialShareReq:
    properties:
      channel:
        type: integer
      content:
        type: string
      link:
        type: string
      share_type:
        type: integer
    type: object
...
  vo.DailyStats:
    properties:
      annual_yield_rate:
        type: number
      day:
        type: string
      game_reward_amount:
        type: integer
      new_address:
        type: integer
      opened_big_reward:
        type: integer
      part_in_address:
        type: integer
      part_in_count:
        type: integer
      reward_amount:
        type: integer
      share_reward_amount:
        type: integer
      single_reward_max:
        type: integer
      single_reward_min:
        type: integer
      staking_amount:
        type: integer
      unopened_big_reward:
        type: integer
      withdraw_amount:
        type: integer
    type: object
...
  vo.ModifyUserProfileReq:
    properties:
      avatar:
        type: integer
      nick_name:
        type: string
    type: object
  vo.Pagination:
    properties:
      curPage:
        type: integer
      items: {}
      totalNum:
        type: integer
    type: object
...
  vo.ProcessSocialShareReq:
    properties:
      accept:
        type: boolean
      auditor:
        type: string
      bonus:
        type: integer
      id:
        type: integer
      reply:
        type: string
    type: object
...
  vo.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  vo.SocialShare:
    properties:
      address:
        type: string
      apply_time:
        type: integer
      auditor:
        type: string
      auditor_address:
        type: string
      bonus:
        type: integer
      channel:
        type: integer
      content:
        type: string
      id:
        type: integer
      link:
        type: string
      nick_name:
        type: string
      reply:
        type: string
      share_type:
        type: integer
      state:
        type: integer
    type: object
...
  vo.UserAuthReq:
    properties:
      address:
        type: string
      nonce:
        type: string
      signed_msg:
        type: string
      timestamp:
        type: string
    type: object
...
  vo.UserProfile:
    properties:
      address:
        type: string
      avatar:
        type: integer
      nick_name:
        type: string
    type: object
  vo.UserYieldHourlyStats:
    properties:
      annually_rate:
        type: number
      balance:
        type: integer
      reward:
        type: integer
      time:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /share:
    get:
      consumes:
      - application/json
      description: get  bonus withdraw ,can filter by address and type
      parameters:
      - description: filter address,if empty will return all user
        in: query
        name: address
        type: string
      - description: filter auditor_address
        in: query
        name: auditor_address
        type: string
      - description: state 0：unProcess 1:processing 2:processed
        in: query
        name: state
        type: integer
      - description: type 1:withdraw 2:common
        in: query
        name: type
        type: integer
      - description: type 1:Gate 2:Weibo 3:Twitter 4:Reddit 5:Facebook
        in: query
        name: channel
        type: integer
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.SocialShare'
                        type: array
                    type: object
              type: object
      summary: GetSocialShares
      tags:
      - share
    post:
      consumes:
      - application/json
      description: add social share
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: addReq
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.AddSocialShareReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: CreateSocialShare
      tags:
      - share
  /share/my:
    get:
      consumes:
      - application/json
      description: get my social shares
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: state 0：unProcess 1:processing 2:processed
        in: query
        name: state
        type: integer
      - description: type 1:withdraw 2:common
        in: query
        name: type
        type: integer
      - description: type 1:Gate 2:Weibo 3:Twitter 4:Reddit 5:Facebook
        in: query
        name: channel
        type: integer
      - description: begin time
        in: query
        name: beginTime
        type: integer
      - description: end time
        in: query
        name: endTime
        type: integer
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/vo.Pagination'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/vo.SocialShare'
                        type: array
                    type: object
              type: object
      summary: GetMySocialShares
      tags:
      - share
  /share/process:
    post:
      consumes:
      - application/json
      description: process social share
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: req
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.ProcessSocialShareReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: SocialShareProcess
      tags:
      - share
  /stats/daily:
    get:
      consumes:
      - application/json
      description: get daily stats
      parameters:
      - description: query day
        in: query
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.DailyStats'
              type: object
      summary: GetDailyStats
      tags:
      - stats
  /stats/daily/list:
    get:
      consumes:
      - application/json
      description: get daily stats list
      parameters:
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.DailyStats'
              type: object
      summary: GetDailyStatsList
      tags:
      - stats
...
  /stats/yield/hourly:
    get:
      consumes:
      - application/json
      description: get user yield hourly list
      parameters:
      - description: page start from 0,default 0
        in: query
        name: page
        type: integer
      - description: size default 20
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.UserYieldHourlyStats'
              type: object
      summary: GetDailyStatsList
      tags:
      - stats
  /user/auth:
    post:
      consumes:
      - application/json
      description: user login and get accessToken
      parameters:
      - description: auth
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.UserAuthReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: auth
      tags:
      - user
  /user/profile:
    get:
      consumes:
      - application/json
      description: get user profile
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.UserProfile'
              type: object
      summary: getUserProfile
      tags:
      - user
    post:
      consumes:
      - application/json
      description: modify user profile
      parameters:
      - description: accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: modifyReq
        in: body
        name: .
        schema:
          $ref: '#/definitions/vo.ModifyUserProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: modifyUserProfile
      tags:
      - user
...
swagger: "2.0"